- Replaced references of IdentityHashtable with Map/IdentityHashMap - Fixed some iterations in uow that were removing from same collection while iterating - Removed unused imports - Fixed some spelling - Removed redundent casts - Removed some unused code

+++ foundation/eclipselink.core.test/src/org/eclipse/persistence/testing/models/inheritance/STI_EmployeePopulator.java
@@ -194,7 +194,7 @@ public class STI_EmployeePopulator {
         try {
             largeProject.setName("Sales Reporting");
             largeProject.setDescription("A reporting application to report on the corporations database through TopLink.");
-            largeProject.setBudget((double)5000);
+            largeProject.setBudget(5000);
         } catch (Exception exception) {
             throw new RuntimeException(exception.toString());
         }
