+++ jpa/eclipselink.jpa.test/src/org/eclipse/persistence/testing/models/jpa/xml/relationships/unidirectional/EmployeePopulator.java
@@ -486,7 +486,7 @@ public class EmployeePopulator {
         try {
             largeProject.setName("Sales Reporting");
             largeProject.setDescription("A reporting application to report on the corporations database through TopLink.");
-            largeProject.setBudget((double)5000);
+            largeProject.setBudget(5000);
             startCalendar.set(1991, 10, 11, 12, 0, 0);
             //largeProject.setMilestoneVersion(new Timestamp(startCalendar.getTime().getTime()));
